BRIEFCASE-82 - Moving to cache utils class.
